added working cli command
Added working cli command, but removed the position due to error now that we aren't creating like 9 different data frames
