+++ base/__init__.py
@@ -1,2 +1,2 @@
 import os
-gdrive = os.path("base/gdrive_key.json")
+gdrive = os.path.join("/Users/a09304858/Documents/GitHub/FanduelResearch/base/gdrive_key.json")
